@@ -9,6 +9,13 @@ import { PurchasesComponent } from './pages/account/components/purchases/purchas
 import { ProfileComponent } from './pages/account/components/profile/profile.component';
 import { KycComponent } from './pages/account/components/kyc/kyc.component';
 import { OrdersComponent } from './pages/account/components/orders/orders.component';
+import { NotFoundComponent } from './pages/not-found/not-found.component';
+import { AccountConfirmationComponent } from './pages/account-confirmation/account-confirmation.component';
+import { TermsAndConditionsComponent } from './pages/terms-and-conditions/terms-and-conditions.component';
+import { ChangePasswordComponent } from './pages/change-password/change-password.component';
+import { authGuard } from './core/guards/auth.guard';
+import { ContactUsComponent } from './pages/contact-us/contact-us.component';
+import { SellerDetailsComponent } from './pages/seller-details/seller-details.component';
 
 export const routes: Routes = [
   {
@@ -19,9 +26,17 @@ export const routes: Routes = [
     path: 'products',
     component: ProductListComponent,
   },
+  { path: 'not-found', component: NotFoundComponent },
+  { path: 'contact', component: ContactUsComponent },
+  {
+    path: 'change-password',
+    pathMatch: 'full',
+    component: ChangePasswordComponent,
+  },
   {
     path: 'account',
     component: AccountComponent,
+    canActivate: [authGuard],
     children: [
       { path: 'dashboard', component: DashboardComponent },
       { path: 'products', component: ProductsComponent },
@@ -29,9 +44,12 @@ export const routes: Routes = [
       { path: 'purchases', component: PurchasesComponent },
       { path: 'profile', component: ProfileComponent },
       { path: 'kyc', component: KycComponent },
-      { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
+      { path: '', redirectTo: 'profile', pathMatch: 'full' },
     ],
   },
-  { path: 'products/:id', component: ProductDetailComponent },
-  { path: '**', redirectTo: '/' },
+  { path: 'account/verify', component: AccountConfirmationComponent },
+  { path: 'terms-and-conditions', component: TermsAndConditionsComponent },
+  { path: 'product/:id', component: ProductDetailComponent },
+  { path: 'shop/:id', component: SellerDetailsComponent },
+  { path: '**', redirectTo: '/not-found' },
 ];
