@@ -1,4 +1,6 @@
 export const environment = {
-  API_URL: 'http://my-prod-url',
+  API_URL: 'https://flexnkentpay.com/api',
+  CLIENT_ID: '2',
+  CLIENT_SECRET: 'c8qdsAItnlv3yXtem9gUq7GW7JRfIw1MyhWtqEIb',
   production: true,
 };
